x86_64 kernel boots and panics with "Hello, World!" message.
The 64-bit kernel panics with a "Hello, World" message. The 32-bit kernel
proceeds and runs normally. The reason for the controlled panic is that I
know the kernel isn't doing the right things yet for x86_64 and will crash.
The work to fix this has yet to be done.

Other changes:
* Stick arch-specific compiler flags in the toolchain files.
* KernelPlatformSupport uses separate source dirs for separate archs.
* Rename get_ebp() to get_frame_pointer(). Because the frame pointer is
  EBP on i386 and RBP on x86_64.
* Moved KERNEL_VIRTUAL_START_ADDR to a header in the Boot/ directory
* 32-bit and 64-bit kernels use a linker script generated from a template.
* Use GAS macros to make the 32-bit assembly boot code easier to read.
* panic() can print to the display even if it happens before the VGA text
  display device has been initialized.

+++ Kernel/include/kernel.hpp
@@ -21,7 +21,7 @@ public:
     // The kernel is the thing that invokes constructors for globals! So, we
     // must do two-phase initialization with a nearly empty constructor and then
     // this method, which performs the real initialization work.
-    void init(multiboot_info_t *mb_info, uint32_t istack);
+    void init(multiboot_info_t *mb_info, uintptr_t istack);
 
     // Enter the event loop until the kernel is done. This may never return.
     void run();
